% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(dataflow, filter = "all", start = NULL, end = NULL)
}
\description{
Returns the dataset for the selected dataflow and filtered information
}
\details{
When writing \code {filter}, write in a vector form and add dot between each filter objects: \code{c("DZA.2..M18T23.")}
}
\examples{
\code{get_data("CAP 2030", filter = c("DZA.2..."), start = 2020, end = 2020)}
}
